+++ packages/react-router-dom/index.tsx
@@ -3,19 +3,22 @@
  * you'll need to update the rollup config for react-router-dom-v5-compat.
  */
 import * as React from "react";
-import {
-  createRoutesFromChildren,
+import type {
   NavigateOptions,
+  RelativeRoutingType,
   RouteObject,
   To,
 } from "react-router";
 import {
   Router,
   createPath,
+  createRoutesFromChildren,
   useHref,
   useLocation,
   useMatch,
+  useMatches,
   useNavigate,
+  useNavigation,
   useResolvedPath,
   UNSAFE_DataRouter as DataRouter,
   UNSAFE_DataRouterProvider as DataRouterProvider,
@@ -101,6 +104,7 @@ export type {
   PathPattern,
   PathRouteProps,
   RedirectFunction,
+  RelativeRoutingType,
   RouteMatch,
   RouteObject,
   RouteProps,
@@ -234,12 +238,8 @@ export function DataBrowserRouter({
   let router = routerSingleton;
 
   return (
-    <DataRouterProvider
-      router={router}
-      basename={basename}
-      fallbackElement={fallbackElement}
-    >
-      <DataRouter />
+    <DataRouterProvider router={router} basename={basename}>
+      <DataRouter fallbackElement={fallbackElement} />
     </DataRouterProvider>
   );
 }
@@ -274,12 +274,8 @@ export function DataHashRouter({
   let router = routerSingleton;
 
   return (
-    <DataRouterProvider
-      router={router}
-      basename={basename}
-      fallbackElement={fallbackElement}
-    >
-      <DataRouter />
+    <DataRouterProvider router={router} basename={basename}>
+      <DataRouter fallbackElement={fallbackElement} />
     </DataRouterProvider>
   );
 }
@@ -400,6 +396,7 @@ export interface LinkProps
   replace?: boolean;
   state?: any;
   resetScroll?: boolean;
+  relative?: RelativeRoutingType;
   to: To;
 }
 
@@ -410,6 +407,7 @@ export const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(
   function LinkWithRef(
     {
       onClick,
+      relative,
       reloadDocument,
       replace,
       state,
@@ -420,12 +418,13 @@ export const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(
     },
     ref
   ) {
-    let href = useHref(to);
+    let href = useHref(to, { relative });
     let internalOnClick = useLinkClickHandler(to, {
       replace,
       state,
       target,
       resetScroll,
+      relative,
     });
     function handleClick(
       event: React.MouseEvent<HTMLAnchorElement, MouseEvent>
@@ -582,6 +581,13 @@ export interface FormProps extends React.FormHTMLAttributes<HTMLFormElement> {
   replace?: boolean;
 
   /**
+   * Determines whether the form action is relative to the route hierarchy or
+   * the pathname.  Use this if you want to opt out of navigating the route
+   * hierarchy and want to instead route based on /-delimited URL segments
+   */
+  relative?: RelativeRoutingType;
+
+  /**
    * A function to call when the form is submitted. If you call
    * `event.preventDefault()` then this form will not do anything.
    */
@@ -627,6 +633,7 @@ const FormImpl = React.forwardRef<HTMLFormElement, FormImplProps>(
       onSubmit,
       fetcherKey,
       routeId,
+      relative,
       ...props
     },
     forwardedRef
@@ -634,7 +641,7 @@ const FormImpl = React.forwardRef<HTMLFormElement, FormImplProps>(
     let submit = useSubmitImpl(fetcherKey, routeId);
     let formMethod: FormMethod =
       method.toLowerCase() === "get" ? "get" : "post";
-    let formAction = useFormAction(action);
+    let formAction = useFormAction(action, { relative });
     let submitHandler: React.FormEventHandler<HTMLFormElement> = (event) => {
       onSubmit && onSubmit(event);
       if (event.defaultPrevented) return;
@@ -643,7 +650,7 @@ const FormImpl = React.forwardRef<HTMLFormElement, FormImplProps>(
       let submitter = (event as unknown as HTMLSubmitEvent).nativeEvent
         .submitter as HTMLFormSubmitter | null;
 
-      submit(submitter || event.currentTarget, { method, replace });
+      submit(submitter || event.currentTarget, { method, replace, relative });
     };
 
     return (
@@ -700,16 +707,18 @@ export function useLinkClickHandler<E extends Element = HTMLAnchorElement>(
     replace: replaceProp,
     state,
     resetScroll,
+    relative,
   }: {
     target?: React.HTMLAttributeAnchorTarget;
     replace?: boolean;
     state?: any;
     resetScroll?: boolean;
+    relative?: RelativeRoutingType;
   } = {}
 ): (event: React.MouseEvent<E, MouseEvent>) => void {
   let navigate = useNavigate();
   let location = useLocation();
-  let path = useResolvedPath(to);
+  let path = useResolvedPath(to, { relative });
 
   return React.useCallback(
     (event: React.MouseEvent<E, MouseEvent>) => {
@@ -717,16 +726,26 @@ export function useLinkClickHandler<E extends Element = HTMLAnchorElement>(
         event.preventDefault();
 
         // If the URL hasn't changed, a regular <a> will do a replace instead of
-        // a push, so do the same here unless the replace prop is explcitly set
+        // a push, so do the same here unless the replace prop is explicitly set
         let replace =
           replaceProp !== undefined
             ? replaceProp
             : createPath(location) === createPath(path);
 
-        navigate(to, { replace, state, resetScroll });
+        navigate(to, { replace, state, resetScroll, relative });
       }
     },
-    [location, navigate, path, replaceProp, state, target, to, resetScroll]
+    [
+      location,
+      navigate,
+      path,
+      replaceProp,
+      state,
+      target,
+      to,
+      resetScroll,
+      relative,
+    ]
   );
 }
 
@@ -864,13 +883,16 @@ function useSubmitImpl(fetcherKey?: string, routeId?: string): SubmitFunction {
   );
 }
 
-export function useFormAction(action?: string): string {
+export function useFormAction(
+  action?: string,
+  { relative }: { relative?: RelativeRoutingType } = {}
+): string {
   let routeContext = React.useContext(RouteContext);
   invariant(routeContext, "useFormAction must be used inside a RouteContext");
 
   let [match] = routeContext.matches.slice(-1);
   let resolvedAction = action ?? ".";
-  let path = useResolvedPath(resolvedAction);
+  let path = useResolvedPath(resolvedAction, { relative });
 
   // Previously we set the default action to ".". The problem with this is that
   // `useResolvedPath(".")` excludes search params and the hash of the resolved
@@ -884,6 +906,15 @@ export function useFormAction(action?: string): string {
     // or hash
     path.search = location.search;
     path.hash = location.hash;
+
+    // When grabbing search params from the URL, remove the automatically
+    // inserted ?index param so we match the useResolvedPath search behavior
+    // which would not include ?index
+    if (match.route.index) {
+      let params = new URLSearchParams(path.search);
+      params.delete("index");
+      path.search = params.toString() ? `?${params.toString()}` : "";
+    }
   }
 
   if ((!action || action === ".") && match.route.index) {
@@ -1008,6 +1039,8 @@ function useScrollRestoration({
   storageKey?: string;
 } = {}) {
   let location = useLocation();
+  let matches = useMatches();
+  let navigation = useNavigation();
   let dataRouterContext = React.useContext(DataRouterContext);
   invariant(
     dataRouterContext,
@@ -1033,10 +1066,8 @@ function useScrollRestoration({
   // Save positions on unload
   useBeforeUnload(
     React.useCallback(() => {
-      if (state?.navigation.state === "idle") {
-        let key =
-          (getKey ? getKey(state.location, state.matches) : null) ||
-          state.location.key;
+      if (navigation.state === "idle") {
+        let key = (getKey ? getKey(location, matches) : null) || location.key;
         savedScrollPositions[key] = window.scrollY;
       }
       sessionStorage.setItem(
@@ -1044,13 +1075,7 @@ function useScrollRestoration({
         JSON.stringify(savedScrollPositions)
       );
       window.history.scrollRestoration = "auto";
-    }, [
-      storageKey,
-      getKey,
-      state.navigation.state,
-      state.location,
-      state.matches,
-    ])
+    }, [storageKey, getKey, navigation.state, location, matches])
   );
 
   // Read in any saved scroll locations