+++ packages/react-router/lib/components.tsx
@@ -22,10 +22,10 @@ import { useSyncExternalStore as useSyncExternalStoreShim } from "./use-sync-ext
 
 import type {
   DataRouteObject,
-  DataRouterContextObject,
   RouteMatch,
   RouteObject,
   Navigator,
+  RelativeRoutingType,
 } from "./context";
 import {
   LocationContext,
@@ -58,6 +58,12 @@ export function _resetModuleScope() {
   routerSingleton = null;
 }
 
+interface DataRouterProviderProps {
+  basename?: string;
+  children?: React.ReactNode;
+  router: RemixRouter;
+}
+
 /**
  * A higher-order component that, given a Remix Router instance. setups the
  * Context's required for data routing
@@ -65,14 +71,8 @@ export function _resetModuleScope() {
 export function DataRouterProvider({
   basename,
   children,
-  fallbackElement,
   router,
-}: {
-  basename?: string;
-  children?: React.ReactNode;
-  fallbackElement?: React.ReactNode;
-  router: RemixRouter;
-}): React.ReactElement {
+}: DataRouterProviderProps): React.ReactElement {
   // Sync router state to our component state to force re-renders
   let state: RouterState = useSyncExternalStoreShim(
     router.subscribe,
@@ -98,29 +98,29 @@ export function DataRouterProvider({
     };
   }, [router]);
 
-  let dataRouterContext: DataRouterContextObject = {
-    router,
-    navigator,
-    static: false,
-    basename: basename || "/",
-  };
-
-  if (!state.initialized) {
-    return <>{fallbackElement}</>;
-  }
-
   return (
-    <DataRouterContext.Provider value={dataRouterContext}>
+    <DataRouterContext.Provider
+      value={{
+        router,
+        navigator,
+        static: false,
+        basename: basename || "/",
+      }}
+    >
       <DataRouterStateContext.Provider value={state} children={children} />
     </DataRouterContext.Provider>
   );
+}
+
+interface DataRouterProps {
+  fallbackElement?: React.ReactNode;
 }
 
 /**
  * A data-aware wrapper for `<Router>` that leverages the Context's provided by
  * `<DataRouterProvider>`
  */
-export function DataRouter() {
+export function DataRouter({ fallbackElement }: DataRouterProps) {
   let dataRouterContext = React.useContext(DataRouterContext);
   invariant(
     dataRouterContext,
@@ -135,7 +135,7 @@ export function DataRouter() {
       navigationType={router.state.historyAction}
       navigator={navigator}
     >
-      <Routes />
+      {router.state.initialized ? <Routes /> : fallbackElement}
     </Router>
   );
 }
@@ -173,12 +173,8 @@ export function DataMemoryRouter({
   let router = routerSingleton;
 
   return (
-    <DataRouterProvider
-      router={router}
-      basename={basename}
-      fallbackElement={fallbackElement}
-    >
-      <DataRouter />
+    <DataRouterProvider router={router} basename={basename}>
+      <DataRouter fallbackElement={fallbackElement} />
     </DataRouterProvider>
   );
 }
@@ -233,6 +229,7 @@ export interface NavigateProps {
   to: To;
   replace?: boolean;
   state?: any;
+  relative?: RelativeRoutingType;
 }
 
 /**
@@ -244,7 +241,12 @@ export interface NavigateProps {
  *
  * @see https://reactrouter.com/docs/en/v6/components/navigate
  */
-export function Navigate({ to, replace, state }: NavigateProps): null {
+export function Navigate({
+  to,
+  replace,
+  state,
+  relative,
+}: NavigateProps): null {
   invariant(
     useInRouterContext(),
     // TODO: This error is probably because they somehow have 2 versions of
@@ -269,7 +271,7 @@ export function Navigate({ to, replace, state }: NavigateProps): null {
     if (dataRouterState && dataRouterState.navigation.state !== "idle") {
       return;
     }
-    navigate(to, { replace, state });
+    navigate(to, { replace, state, relative });
   });
 
   return null;